Temperature controller modified

+++ Marlin/Marlin.h
@@ -51,8 +51,8 @@ void manage_inactivity(byte debug);
 void manage_heater();
 //int temp2analogu(int celsius, const short table[][2], int numtemps);
 //float analog2tempu(int raw, const short table[][2], int numtemps);
-float temp2analog(int celsius);
-float temp2analogBed(int celsius);
+int temp2analog(float celsius);
+int temp2analogBed(float celsius);
 float analog2temp(int raw);
 float analog2tempBed(int raw);
